pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chemagent"
version = "0.1.0"
description = "Chemistry Enhancement Package for Claude Code and Gemini CLI"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "ChemAgent Team", email = "contact@chemagent.ai"},
]
keywords = ["chemistry", "ai", "agent", "llm", "rdkit", "drug-discovery"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "asyncio>=3.4.3",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.0.0",
    "rdkit>=2023.9.1",
    "pubchempy>=1.0.4",
    "biopython>=1.81",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "ipython>=8.14.0",
    "jupyter>=1.0.0",
]

mcp = [
    "mcp>=0.1.0",
    "websockets>=11.0",
    "jsonrpc>=3.0",
]

visualization = [
    "py3dmol>=2.0.0",
    "nglview>=3.0.0",
    "ipywidgets>=8.0.0",
]

quantum = [
    "psi4>=1.8.0",
    "qcengine>=0.26.0",
    "qcelemental>=0.25.0",
]

[project.scripts]
chemagent = "chemagent_install:main"

[project.urls]
Homepage = "https://github.com/yourusername/chemagent"
Documentation = "https://chemagent.readthedocs.io"
Repository = "https://github.com/yourusername/chemagent"
Issues = "https://github.com/yourusername/chemagent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["chemagent*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
